@@ -18,8 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.6.2 <4.0.0"
-  flutter: ">=3.27.4"
+  sdk: ">=3.7.2 <4.0.0"
+  flutter: ">=3.29.3"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
