@@ -58,7 +58,7 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   CustomPolygonLayersExample? _customPolygonLayersExample;
-  late final AppLifecycleListener _listener;
+  late final AppLifecycleListener _appLifecycleListener;
 
   @override
   Widget build(BuildContext context) {
@@ -129,7 +129,7 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
-    _listener = AppLifecycleListener(
+    _appLifecycleListener = AppLifecycleListener(
       onDetach: () =>
       // Sometimes Flutter may not reliably call dispose(),
       // therefore it is recommended to dispose the HERE SDK
@@ -150,7 +150,7 @@ class _MyAppState extends State<MyApp> {
     _customPolygonLayersExample?.onDestroy();
     await SDKNativeEngine.sharedInstance?.dispose();
     SdkContext.release();
-    _listener.dispose();
+    _appLifecycleListener.dispose();
   }
 
   // A helper method to add a button on top of the HERE map.
