@@ -57,7 +57,7 @@ class EVRoutingApp extends StatefulWidget {
 
 class _EVRoutingAppState extends State<EVRoutingApp> {
   EVRoutingExample? _evRoutingExample;
-  late final AppLifecycleListener _listener;
+  late final AppLifecycleListener _appLifecycleListener;
 
   @override
   Widget build(BuildContext context) {
@@ -106,7 +106,7 @@ class _EVRoutingAppState extends State<EVRoutingApp> {
   @override
   void initState() {
     super.initState();
-    _listener = AppLifecycleListener(
+    _appLifecycleListener = AppLifecycleListener(
       onDetach: () =>
       // Sometimes Flutter may not reliably call dispose(),
       // therefore it is recommended to dispose the HERE SDK
@@ -126,7 +126,7 @@ class _EVRoutingAppState extends State<EVRoutingApp> {
     // Free HERE SDK resources before the application shuts down.
     await SDKNativeEngine.sharedInstance?.dispose();
     SdkContext.release();
-    _listener.dispose();
+    _appLifecycleListener.dispose();
   }
 
   // A helper method to add a button on top of the HERE map.
