@@ -62,7 +62,7 @@ class _MyAppState extends State<MyApp> {
   MapItemsExample? _mapItemsExample;
   MapObjectsExample? _mapObjectsExample;
   MapViewPinsExample? _mapViewPinsExample;
-  late final AppLifecycleListener _listener;
+  late final AppLifecycleListener _appLifecycleListener;
 
   @override
   Widget build(BuildContext context) {
@@ -269,7 +269,7 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
-    _listener = AppLifecycleListener(
+    _appLifecycleListener = AppLifecycleListener(
       onDetach: () =>
       // Sometimes Flutter may not reliably call dispose(),
       // therefore it is recommended to dispose the HERE SDK
@@ -289,7 +289,7 @@ class _MyAppState extends State<MyApp> {
     // Free HERE SDK resources before the application shuts down.
     await SDKNativeEngine.sharedInstance?.dispose();
     SdkContext.release();
-    _listener.dispose();
+    _appLifecycleListener.dispose();
   }
 
   // A helper method to add a button on top of the HERE map.
