@@ -59,7 +59,7 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   late HereMapController? _hereMapController;
   late SearchEngine? _searchEngine;
-  late final AppLifecycleListener _listener;
+  late final AppLifecycleListener _appLifecycleListener;
   late MapScheme _currentMapScheme;
   late IconProvider? _iconProvider;
 
@@ -229,7 +229,7 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
-    _listener = AppLifecycleListener(
+    _appLifecycleListener = AppLifecycleListener(
       onDetach: () =>
       // Sometimes Flutter may not reliably call dispose(),
       // therefore it is recommended to dispose the HERE SDK
@@ -249,7 +249,7 @@ class _MyAppState extends State<MyApp> {
     // Free HERE SDK resources before the application shuts down.
     await SDKNativeEngine.sharedInstance?.dispose();
     SdkContext.release();
-    _listener.dispose();
+    _appLifecycleListener.dispose();
   }
 
   // A helper method to show a dialog.
